调整代码

+++ spring-support-mybatis-starter/src/main/java/com/chua/starter/mybatis/controller/AbstractBaseController.java
@@ -5,12 +5,15 @@ import com.baomidou.mybatisplus.extension.plugins.pagination.Page;
 import com.baomidou.mybatisplus.extension.service.IService;
 import com.chua.starter.common.support.result.ResultData;
 import com.chua.starter.mybatis.entity.DelegatePage;
+import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
+import springfox.documentation.annotations.ApiIgnore;
 
 import javax.annotation.Resource;
+import javax.validation.Valid;
 
 /**
  * 超类
@@ -27,7 +30,10 @@ public abstract class AbstractBaseController<S extends IService<T>, T> {
      * @return 分页结果
      */
     @GetMapping("page")
-    public ResultData<Page<T>> page(DelegatePage<T> page, T entity) {
+    public ResultData<Page<T>> page(DelegatePage<T> page, @Valid T entity, BindingResult bindingResult) {
+        if (bindingResult.hasErrors()) {
+            return ResultData.failure(1, bindingResult.getAllErrors().get(0).getDefaultMessage());
+        }
         return ResultData.success(getService().page(page.createPage(), Wrappers.lambdaQuery(entity)));
     }
 
@@ -39,6 +45,9 @@ public abstract class AbstractBaseController<S extends IService<T>, T> {
      */
     @GetMapping("delete/{id}")
     public ResultData<Boolean> delete(@PathVariable("id") String id) {
+        if (null == id) {
+            return ResultData.failure(1, "主键不能为空");
+        }
         return ResultData.success(getService().removeById(id));
     }
 
@@ -49,7 +58,10 @@ public abstract class AbstractBaseController<S extends IService<T>, T> {
      * @return 分页结果
      */
     @PostMapping("update")
-    public ResultData<Boolean> updateById(@RequestBody T t) {
+    public ResultData<Boolean> updateById(@Valid @RequestBody T t, BindingResult bindingResult) {
+        if (bindingResult.hasErrors()) {
+            return ResultData.failure(1, bindingResult.getAllErrors().get(0).getDefaultMessage());
+        }
         return ResultData.success(getService().updateById(t));
     }
 
@@ -60,7 +72,10 @@ public abstract class AbstractBaseController<S extends IService<T>, T> {
      * @return 分页结果
      */
     @PostMapping("save")
-    public ResultData<Boolean> save(@RequestBody T t) {
+    public ResultData<Boolean> save(@Valid @RequestBody T t, BindingResult bindingResult) {
+        if (bindingResult.hasErrors()) {
+            return ResultData.failure(1, bindingResult.getAllErrors().get(0).getDefaultMessage());
+        }
         return ResultData.success(getService().save(t));
     }
