+++ spring-support-mybatis-starter/src/main/java/com/chua/starter/mybatis/controller/AbstractSwaggerController.java
@@ -7,7 +7,11 @@ import com.chua.starter.mybatis.entity.RequestPage;
 import com.chua.starter.mybatis.entity.ResultPage;
 import io.swagger.annotations.ApiOperation;
 import io.swagger.annotations.ApiParam;
+import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.*;
+import springfox.documentation.annotations.ApiIgnore;
+
+import javax.validation.Valid;
 
 /**
  * 超类
@@ -24,8 +28,11 @@ public abstract class AbstractSwaggerController<S extends IService<T>, T> {
      */
     @ApiOperation("分页查询基础数据")
     @GetMapping("page")
-    public ResultData<ResultPage<T>> page(RequestPage<T> page, T entity) {
-        return ResultData.success(ResultPage.copyList(getService().page(page.createPage(), Wrappers.lambdaQuery(entity))));
+    public ResultData<ResultPage<T>> page(@Valid RequestPage<T> page, T entity, @ApiIgnore BindingResult bindingResult) {
+        if(bindingResult.hasErrors()) {
+            return ResultData.failure(1, bindingResult.getAllErrors().get(0).getDefaultMessage());
+        }
+        return ResultData.success(ResultPage.copy(getService().page(page.createPage(), Wrappers.lambdaQuery(entity))));
     }
 
     /**
@@ -37,6 +44,10 @@ public abstract class AbstractSwaggerController<S extends IService<T>, T> {
     @ApiOperation("根据主键删除数据")
     @GetMapping("delete/{id}")
     public ResultData<Boolean> delete(@ApiParam("主键") @PathVariable("id") String id) {
+        if(null == id) {
+            return ResultData.failure(1,  "主键不能为空");
+        }
+
         return ResultData.success(getService().removeById(id));
     }
 
@@ -48,7 +59,10 @@ public abstract class AbstractSwaggerController<S extends IService<T>, T> {
      */
     @ApiOperation("根据主键更新数据")
     @PostMapping("update")
-    public ResultData<Boolean> updateById(@RequestBody T t) {
+    public ResultData<Boolean> updateById(@Valid @RequestBody T t , @ApiIgnore BindingResult bindingResult) {
+        if(bindingResult.hasErrors()) {
+            return ResultData.failure(1, bindingResult.getAllErrors().get(0).getDefaultMessage());
+        }
         return ResultData.success(getService().updateById(t));
     }
 
@@ -60,7 +74,10 @@ public abstract class AbstractSwaggerController<S extends IService<T>, T> {
      */
     @ApiOperation("上报数据")
     @PostMapping("save")
-    public ResultData<Boolean> save(@RequestBody T t) {
+    public ResultData<Boolean> save(@Valid @RequestBody T t, @ApiIgnore BindingResult bindingResult) {
+        if(bindingResult.hasErrors()) {
+            return ResultData.failure(1, bindingResult.getAllErrors().get(0).getDefaultMessage());
+        }
         return ResultData.success(getService().save(t));
     }
